+++ trunk/src/edr/src/plugins/video/VideoVisualizer.h
@@ -100,14 +100,14 @@ private:
         }
 
         //! \return Dlugosc kanalu w sekundach
-        virtual float getLength() const
+        virtual double getLength() const
         {
             return visualizer->stream->getDuration();
         }
 
         //! Czas zawiera siê miêdzy 0 a getLength()
         //! \param time Aktualny, lokalny czas kanalu w sekundach
-        virtual void setTime(float time)
+        virtual void setTime(double time)
         {
             osg::const_pointer_cast<VideoStream>(visualizer->stream)->setTime(time);
             visualizer->streamImage = visualizer->stream->getImage(vidlib::PixelFormatRGB24);
